Removed informal tests from State.py and deleted __init__ method.
There's no (simple) way to tell the difference between a "state"
function and any other function passed into the constructor. If a
non-state function is passed in, then the applicative instance doesn't
function correctly. So, only state function should be passed directly to
the constructor, all other functions and/or value must be injected into
State using the 'unit' method.